Ensure PTT button reflects ASR state
Before this change, the push-to-talk button was always enabled and
did not reflect the ASR state. This made it difficult for the user
to understand if the PTT could be clicked, and what a click would mean.
It was also impossible to request the ASR to stop listening.

After this change, the PTT button reflects the ASR state. For example,
the button says "start listen" or "stop listen" depending on the
state. The change makes it possible to request the ASR to stop
listening.

Additionally, the layout is improved so that backend status and
ASR status do not intersect in an undesired way.