Use threads appropriately
Previously, the testapplication assumed that all callbacks
from TdmConnector listeners happened on the UI thread,
although that is not necessarily the case.

This change ensures that all GUI modifications, typically
following from the listener callbacks, take place on the UI
thread.

Furthermore, TdmConnector.connect() and .disconnect()
previously had to be called from the main thread. With a
recent update to the TDM frontend library, this is no longer
the case either.

This change simplifies the code, so that these TdmConnector
methods are not necessarily called from the main thread
anymore.