Upgrade to AFL 9.0.0
Before this change, the supported AFL version was 7.0.0, using the
websocket channel. With TDM 8.0.0 however, it was replaced by an HTTP
API.

This change upgrades AFL to support version 2.0 of the HTTP API,
introduced in TDM 9.0.0.

This removes interpreted utterance, 'IU>', and active DDD, 'D>',
information from the layout, since the information is no longer
forwarded explicitly by AFL's IEventListener.