Only attempt to call when the action is invoked
Before this change, the app wouldn't check if it had received the
'call' action before starting to look at parameters that were only
expected for the 'call' action. This would crash the app with a
NullPointerException.

This change ensures that a call is only attempted when the 'call'
action has been received.